Add standalone example
A separate test.html file reproduces the initial example which was based
on Vite and Svelte. This example works in Firefox as intended. Based on
this the Svelte example was modified to show that it does work in
Firefox as well.

It seems like the SVG is treated somewhat differently in Firefox and
thereby more easily removed if not immediately used.